+++ Source/SA/Render/RHI/Buffer/RHID12Buffer.cpp
@@ -8,7 +8,7 @@
 
 namespace SA::RND::RHI
 {
-	void D12Buffer::Create(const Device* _device, uint32_t _size, BufferUsage _usage, const void* _src)
+	void D12Buffer::Create(const Device* _device, uint64_t _size, BufferUsage _usage, const void* _src)
 	{
 		mHandle.Create(_device->API_DirectX12(),
 			_size,
